# perf: run parse_compilation_units once