--- a/src/dune_rules/dep_graph.ml
+++ b/src/dune_rules/dep_graph.ml
@@ -3,20 +3,20 @@ open Action_builder.O
 
 type t =
   { dir : Path.Build.t
-  ; per_module : Module.t list Action_builder.t Module.Obj_map.t
+  ; per_module : Module.t list Action_builder.t Module_name.Unique.Map.t
   }
 
 let make ~dir ~per_module = { dir; per_module }
 
 let deps_of t (m : Module.t) =
-  match Module.Obj_map.find t.per_module m with
+  match Module_name.Unique.Map.find t.per_module (Module.obj_name m) with
   | Some x -> x
   | None ->
     Code_error.raise "Ocamldep.Dep_graph.deps_of"
       [ ("dir", Path.Build.to_dyn t.dir)
       ; ( "modules"
         , Dyn.(list Module_name.Unique.to_dyn)
-            (Module.Obj_map.keys t.per_module |> List.map ~f:Module.obj_name) )
+            (Module_name.Unique.Map.keys t.per_module) )
       ; ("m", Module.to_dyn m)
       ]
 
@@ -24,8 +24,8 @@ module Top_closure = Top_closure.Make (Module_name.Unique.Set) (Action_builder)
 
 let top_closed t modules =
   let+ res =
-    Top_closure.top_closure modules ~key:Module.obj_name
-      ~deps:(Module.Obj_map.find_exn t.per_module)
+    Top_closure.top_closure modules ~key:Module.obj_name ~deps:(fun m ->
+        Module_name.Unique.Map.find_exn t.per_module (Module.obj_name m))
   in
   match res with
   | Ok modules -> modules
@@ -58,7 +58,9 @@ let top_closed_implementations t modules =
 
 let dummy (m : Module.t) =
   { dir = Path.Build.root
-  ; per_module = Module.Obj_map.singleton m (Action_builder.return [])
+  ; per_module =
+      Module_name.Unique.Map.singleton (Module.obj_name m)
+        (Action_builder.return [])
   }
 
 module Ml_kind = struct