--- a/src/dune_rules/dep_rules.ml
+++ b/src/dune_rules/dep_rules.ml
@@ -3,11 +3,14 @@ open Memo.O
 open Ocamldep.Modules_data
 
 let transitive_deps_contents modules =
-  List.map modules ~f:(fun m -> Module_name.to_string (Module.name m))
+  List.map modules ~f:(fun m ->
+      Module_name.to_string (Module.name (Modules.Sourced_module.to_module m)))
   |> String.concat ~sep:"\n"
 
 let ooi_deps { vimpl; sctx; dir; obj_dir; modules = _; stdlib = _; sandbox = _ }
-    ~dune_version ~vlib_obj_map ~(ml_kind : Ml_kind.t) (m : Module.t) =
+    ~dune_version ~vlib_obj_map ~(ml_kind : Ml_kind.t)
+    (sourced_module : Modules.Sourced_module.t) =
+  let m = Modules.Sourced_module.to_module sourced_module in
   let cm_kind =
     match ml_kind with
     | Intf -> Cm_kind.Cmi
@@ -54,14 +57,16 @@ let deps_of_module ({ modules; _ } as md) ~ml_kind m =
     List.singleton interface_module |> Action_builder.return |> Memo.return
   | _ -> (
     let+ deps = Ocamldep.deps_of md ~ml_kind m in
+    let open Action_builder.O in
+    let+ deps = deps in
+    let deps = List.map ~f:Modules.Sourced_module.to_module deps in
     match Modules.alias_for modules m with
     | [] -> deps
-    | aliases ->
-      let open Action_builder.O in
-      let+ deps = deps in
-      aliases @ deps)
+    | aliases -> aliases @ deps)
 
-let deps_of_vlib_module ({ obj_dir; vimpl; dir; sctx; _ } as md) ~ml_kind m =
+let deps_of_vlib_module ({ obj_dir; vimpl; dir; sctx; _ } as md) ~ml_kind
+    sourced_module =
+  let m = Modules.Sourced_module.to_module sourced_module in
   let vimpl = Option.value_exn vimpl in
   let vlib = Vimpl.vlib vimpl in
   match Lib.Local.of_lib vlib with
@@ -71,7 +76,13 @@ let deps_of_vlib_module ({ obj_dir; vimpl; dir; sctx; _ } as md) ~ml_kind m =
       let impl = Vimpl.impl vimpl in
       Dune_project.dune_version impl.project
     in
-    ooi_deps md ~dune_version ~vlib_obj_map ~ml_kind m
+    let open Memo.O in
+    let+ deps =
+      ooi_deps md ~dune_version ~vlib_obj_map ~ml_kind sourced_module
+    in
+    let open Action_builder.O in
+    let+ deps = deps in
+    List.map ~f:Modules.Sourced_module.to_module deps
   | Some lib ->
     let modules = Vimpl.vlib_modules vimpl in
     let info = Lib.Local.info lib in
@@ -83,7 +94,11 @@ let deps_of_vlib_module ({ obj_dir; vimpl; dir; sctx; _ } as md) ~ml_kind m =
     let+ () =
       Super_context.add_rule sctx ~dir (Action_builder.symlink ~src ~dst)
     in
-    Ocamldep.read_deps_of ~obj_dir:vlib_obj_dir ~modules ~ml_kind m
+    let open Action_builder.O in
+    let+ deps =
+      Ocamldep.read_deps_of ~obj_dir:vlib_obj_dir ~modules ~ml_kind m
+    in
+    List.map ~f:Modules.Sourced_module.to_module deps
 
 let rec deps_of md ~ml_kind (m : Modules.Sourced_module.t) =
   let is_alias =
@@ -96,12 +111,13 @@ let rec deps_of md ~ml_kind (m : Modules.Sourced_module.t) =
   in
   if is_alias then Memo.return (Action_builder.return [])
   else
-    let skip_if_source_absent f m =
-      if Module.has m ~ml_kind then f m
+    let skip_if_source_absent f sourced_module =
+      let m = Modules.Sourced_module.to_module m in
+      if Module.has m ~ml_kind then f sourced_module
       else Memo.return (Action_builder.return [])
     in
     match m with
-    | Imported_from_vlib m ->
+    | Imported_from_vlib _ ->
       skip_if_source_absent (deps_of_vlib_module md ~ml_kind) m
     | Normal m -> skip_if_source_absent (deps_of_module md ~ml_kind) m
     | Impl_of_virtual_module impl_or_vlib -> (