--- a/src/dune_rules/vimpl.mli
+++ b/src/dune_rules/vimpl.mli
@@ -28,4 +28,4 @@ val vlib_stubs_o_files : t option -> Path.t list
 
 val impl_cm_kind : t -> Cm_kind.t
 
-val vlib_obj_map : t -> Module.t Module_name.Unique.Map.t
+val vlib_obj_map : t -> Modules.Sourced_module.t Module_name.Unique.Map.t