`parse_compilation_units` can be expensive when there are a lot of modules in the map. This PR makes sure it only runs once per lib (or other stanzas).

Some measurements for a few actions in Ahrefs repo. All actions are "noops" (the command was run previously successfully so no user actions should be executed):

| Command | `main`  | This branch |
| ------------- | ------------- |  ------------- |
| `dune exec gen_esgg`  | `real	0m10.121s, user	0m9.677s, sys	0m0.440s` | `real	0m1.372s, user	0m1.157s, sys	0m0.216s`  |
| `dune build @fe.all`  | `real	0m32.039s, user	0m30.391s, sys	0m1.685s` | `real	0m12.716s, user	0m11.615s, sys	0m1.119s`  |
| `dune build @gen_dune`  | `real	0m25.436s, user	0m24.253s, sys	0m1.223s` | `real	0m5.407s, user	0m4.908s, sys	0m0.504s`  |

### TODO

- [x] Create public benchmark to test this PR (#7189, #7198)