@rgrinberg 2023-02-26
Fix looks right. I'm wondering if we could simplify things even further though. How about saving the object map as part of `Modules.t`? This would remove a lot of the staging as well.