@jchavarri 2023-02-28
@rgrinberg how do you think the types should be modified in order to move in that direction? Something like this? 

```
type t =
  { obj_map : Module.t Module_name.Unique.Map.t
  ; modules : modules
  }

and modules =
  | Singleton of Module.t
  | Unwrapped of Unwrapped.t
  | Wrapped of Wrapped.t
  | Impl of impl
  | Stdlib of Stdlib.t

and impl =
  { impl : t
  ; vlib : t
  }
```