@jchavarri 2023-02-28
Another thing I wonder about the idea of saving the object map as part of `Modules.t` is if there can be cases where this object map would be created unnecessarily, in a wasteful way 🤔