@rgrinberg 2023-02-28
> how do you think the types should be modified in order to move in that direction? Something like this?

Yeah that sounds good.

> Another thing I wonder about the idea of saving the object map as part of Modules.t is if there can be cases where this object map would be created unnecessarily, in a wasteful way 🤔

You could make it lazy to avoid that.