@jchavarri 2023-03-02
@rgrinberg thanks for the review, I think I tackled all your comments, the code looks better now.

If it looks good to merge, before doing so please take a look at #7198. I would like to merge that benchmark PR first, so that we can make measurements compare and understand better the impact in warm builds, thanks!