@rgrinberg 2023-03-03
I simplified things a little further. We don't need Obj_map at all.